Switch to using named pipes for test communication (microsoft/vscode-python#22390)
Resolves https://github.com/microsoft/vscode-python/issues/22177.
Switches to using named pipes as communication between extension and
test run subprocesses.

---------

+++ extensions/positron-python/src/test/testing/common/testingAdapter.test.ts
@@ -8,13 +8,11 @@ import * as assert from 'assert';
 import * as fs from 'fs';
 import { PytestTestDiscoveryAdapter } from '../../../client/testing/testController/pytest/pytestDiscoveryAdapter';
 import { ITestController, ITestResultResolver } from '../../../client/testing/testController/common/types';
-import { PythonTestServer } from '../../../client/testing/testController/common/server';
 import { IPythonExecutionFactory } from '../../../client/common/process/types';
-import { ITestDebugLauncher } from '../../../client/testing/common/types';
 import { IConfigurationService, ITestOutputChannel } from '../../../client/common/types';
 import { IServiceContainer } from '../../../client/ioc/types';
 import { EXTENSION_ROOT_DIR_FOR_TESTS, initialize } from '../../initialize';
-import { traceLog } from '../../../client/logging';
+import { traceError, traceLog } from '../../../client/logging';
 import { PytestTestExecutionAdapter } from '../../../client/testing/testController/pytest/pytestExecutionAdapter';
 import { UnittestTestDiscoveryAdapter } from '../../../client/testing/testController/unittest/testDiscoveryAdapter';
 import { UnittestTestExecutionAdapter } from '../../../client/testing/testController/unittest/testExecutionAdapter';
@@ -25,9 +23,7 @@ import { IEnvironmentVariablesProvider } from '../../../client/common/variables/
 
 suite('End to End Tests: test adapters', () => {
     let resultResolver: ITestResultResolver;
-    let pythonTestServer: PythonTestServer;
     let pythonExecFactory: IPythonExecutionFactory;
-    let debugLauncher: ITestDebugLauncher;
     let configService: IConfigurationService;
     let serviceContainer: IServiceContainer;
     let envVarsService: IEnvironmentVariablesProvider;
@@ -72,26 +68,27 @@ suite('End to End Tests: test adapters', () => {
         // create symlink for specific symlink test
         const target = rootPathSmallWorkspace;
         const dest = rootPathDiscoverySymlink;
-        fs.symlink(target, dest, 'dir', (err) => {
-            if (err) {
-                console.error(err);
-            } else {
-                console.log('Symlink created successfully for end to end tests.');
-            }
-        });
+        try {
+            fs.symlink(target, dest, 'dir', (err) => {
+                if (err) {
+                    traceError(err);
+                } else {
+                    traceLog('Symlink created successfully for end to end tests.');
+                }
+            });
+        } catch (err) {
+            traceError(err);
+        }
     });
 
     setup(async () => {
         // create objects that were injected
         configService = serviceContainer.get<IConfigurationService>(IConfigurationService);
         pythonExecFactory = serviceContainer.get<IPythonExecutionFactory>(IPythonExecutionFactory);
-        debugLauncher = serviceContainer.get<ITestDebugLauncher>(ITestDebugLauncher);
         testController = serviceContainer.get<TestController>(ITestController);
         envVarsService = serviceContainer.get<IEnvironmentVariablesProvider>(IEnvironmentVariablesProvider);
 
         // create objects that were not injected
-        pythonTestServer = new PythonTestServer(pythonExecFactory, debugLauncher);
-        await pythonTestServer.serverReady();
 
         testOutputChannel = typeMoq.Mock.ofType<ITestOutputChannel>();
         testOutputChannel
@@ -111,19 +108,20 @@ suite('End to End Tests: test adapters', () => {
                 // Whatever you need to return
             });
     });
-    teardown(async () => {
-        pythonTestServer.dispose();
-    });
     suiteTeardown(async () => {
         // remove symlink
         const dest = rootPathDiscoverySymlink;
-        fs.unlink(dest, (err) => {
-            if (err) {
-                console.error(err);
-            } else {
-                console.log('Symlink removed successfully after tests.');
-            }
-        });
+        if (fs.existsSync(dest)) {
+            fs.unlink(dest, (err) => {
+                if (err) {
+                    traceError(err);
+                } else {
+                    traceLog('Symlink removed successfully after tests.');
+                }
+            });
+        } else {
+            traceLog('Symlink was not found to remove after tests, exiting successfully');
+        }
     });
     test('unittest discovery adapter small workspace', async () => {
         // result resolver and saved data for assertions
@@ -136,6 +134,7 @@ suite('End to End Tests: test adapters', () => {
         workspaceUri = Uri.parse(rootPathSmallWorkspace);
         resultResolver = new PythonResultResolver(testController, unittestProvider, workspaceUri);
         let callCount = 0;
+        // const deferredTillEOT = createTestingDeferred();
         resultResolver._resolveDiscovery = async (payload, _token?) => {
             traceLog(`resolveDiscovery ${payload}`);
             callCount = callCount + 1;
@@ -149,14 +148,13 @@ suite('End to End Tests: test adapters', () => {
 
         // run unittest discovery
         const discoveryAdapter = new UnittestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
             envVarsService,
         );
 
-        await discoveryAdapter.discoverTests(workspaceUri).finally(() => {
+        await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
 
             // 1. Check the status is "success"
@@ -196,14 +194,13 @@ suite('End to End Tests: test adapters', () => {
         configService.getSettings(workspaceUri).testing.unittestArgs = ['-s', '.', '-p', '*test*.py'];
         // run discovery
         const discoveryAdapter = new UnittestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
             envVarsService,
         );
 
-        await discoveryAdapter.discoverTests(workspaceUri).finally(() => {
+        await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // 1. Check the status is "success"
             assert.strictEqual(
                 actualData.status,
@@ -226,25 +223,23 @@ suite('End to End Tests: test adapters', () => {
             status: 'success' | 'error';
             error?: string[];
         };
+        // set workspace to test workspace folder
+        workspaceUri = Uri.parse(rootPathSmallWorkspace);
         resultResolver = new PythonResultResolver(testController, pytestProvider, workspaceUri);
         let callCount = 0;
         resultResolver._resolveDiscovery = async (payload, _token?) => {
-            traceLog(`resolveDiscovery ${payload}`);
             callCount = callCount + 1;
             actualData = payload;
             return Promise.resolve();
         };
         // run pytest discovery
         const discoveryAdapter = new PytestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
             envVarsService,
         );
 
-        // set workspace to test workspace folder
-        workspaceUri = Uri.parse(rootPathSmallWorkspace);
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
 
@@ -287,12 +282,12 @@ suite('End to End Tests: test adapters', () => {
         };
         // run pytest discovery
         const discoveryAdapter = new PytestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
             envVarsService,
         );
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
@@ -363,7 +358,6 @@ suite('End to End Tests: test adapters', () => {
         };
         // run pytest discovery
         const discoveryAdapter = new PytestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -372,6 +366,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathLargeWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
@@ -417,7 +412,6 @@ suite('End to End Tests: test adapters', () => {
         configService.getSettings(workspaceUri).testing.unittestArgs = ['-s', '.', '-p', '*test*.py'];
         // run execution
         const executionAdapter = new UnittestTestExecutionAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -428,9 +422,9 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
         let collectedOutput = '';
         testRun
@@ -441,7 +435,13 @@ suite('End to End Tests: test adapters', () => {
             })
             .returns(() => false);
         await executionAdapter
-            .runTests(workspaceUri, ['test_simple.SimpleClass.test_simple_unit'], false, testRun.object)
+            .runTests(
+                workspaceUri,
+                ['test_simple.SimpleClass.test_simple_unit'],
+                false,
+                testRun.object,
+                pythonExecFactory,
+            )
             .finally(() => {
                 // verify that the _resolveExecution was called once per test
                 assert.strictEqual(callCount, 1, 'Expected _resolveExecution to be called once');
@@ -492,7 +492,6 @@ suite('End to End Tests: test adapters', () => {
 
         // run unittest execution
         const executionAdapter = new UnittestTestExecutionAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -503,9 +502,9 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
         let collectedOutput = '';
         testRun
@@ -516,7 +515,13 @@ suite('End to End Tests: test adapters', () => {
             })
             .returns(() => false);
         await executionAdapter
-            .runTests(workspaceUri, ['test_parameterized_subtest.NumbersTest.test_even'], false, testRun.object)
+            .runTests(
+                workspaceUri,
+                ['test_parameterized_subtest.NumbersTest.test_even'],
+                false,
+                testRun.object,
+                pythonExecFactory,
+            )
             .then(() => {
                 // verify that the _resolveExecution was called once per test
                 assert.strictEqual(callCount, 2000, 'Expected _resolveExecution to be called once');
@@ -558,10 +563,10 @@ suite('End to End Tests: test adapters', () => {
         };
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathSmallWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         // run pytest execution
         const executionAdapter = new PytestTestExecutionAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -572,9 +577,9 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
         let collectedOutput = '';
         testRun
@@ -648,6 +653,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathLargeWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         // generate list of test_ids
         const testIds: string[] = [];
@@ -658,7 +664,6 @@ suite('End to End Tests: test adapters', () => {
 
         // run pytest execution
         const executionAdapter = new PytestTestExecutionAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -669,9 +674,9 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
         let collectedOutput = '';
         testRun
@@ -728,9 +733,9 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathDiscoveryErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.unittestArgs = ['-s', '.', '-p', '*test*.py'];
 
         const discoveryAdapter = new UnittestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -741,11 +746,11 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
-        await discoveryAdapter.discoverTests(workspaceUri).finally(() => {
+        await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             assert.strictEqual(callCount, 1, 'Expected _resolveDiscovery to be called once');
             assert.strictEqual(failureOccurred, false, failureMsg);
         });
@@ -790,7 +795,6 @@ suite('End to End Tests: test adapters', () => {
         };
         // run pytest discovery
         const discoveryAdapter = new PytestTestDiscoveryAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -799,6 +803,8 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathDiscoveryErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
+
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
             assert.ok(
@@ -860,10 +866,10 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.unittestArgs = ['-s', '.', '-p', '*test*.py'];
 
         // run pytest execution
         const executionAdapter = new UnittestTestExecutionAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -874,11 +880,11 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
-        await executionAdapter.runTests(workspaceUri, testIds, false, testRun.object).finally(() => {
+        await executionAdapter.runTests(workspaceUri, testIds, false, testRun.object, pythonExecFactory).finally(() => {
             assert.strictEqual(callCount, 1, 'Expected _resolveExecution to be called once');
             assert.strictEqual(failureOccurred, false, failureMsg);
         });
@@ -913,7 +919,7 @@ suite('End to End Tests: test adapters', () => {
                 failureMsg = err ? (err as Error).toString() : '';
                 failureOccurred = true;
             }
-            return Promise.resolve();
+            // return Promise.resolve();
         };
 
         const testId = `${rootPathErrorWorkspace}/test_seg_fault.py::TestSegmentationFault::test_segfault`;
@@ -921,10 +927,10 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         // run pytest execution
         const executionAdapter = new PytestTestExecutionAdapter(
-            pythonTestServer,
             configService,
             testOutputChannel.object,
             resultResolver,
@@ -935,9 +941,9 @@ suite('End to End Tests: test adapters', () => {
             .setup((t) => t.token)
             .returns(
                 () =>
-                    ({
-                        onCancellationRequested: () => undefined,
-                    } as any),
+                ({
+                    onCancellationRequested: () => undefined,
+                } as any),
             );
         await executionAdapter.runTests(workspaceUri, testIds, false, testRun.object, pythonExecFactory).finally(() => {
             assert.strictEqual(callCount, 1, 'Expected _resolveExecution to be called once');
